Rename and refactor classes
