+++ app/src/main/java/fr/free/nrw/commons/bookmarks/pictures/BookmarkPictureContentProvider.java
@@ -1,7 +1,6 @@
-package fr.free.nrw.commons.bookmarks;
+package fr.free.nrw.commons.bookmarks.pictures;
 
 import android.content.ContentValues;
-import android.content.UriMatcher;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteQueryBuilder;
@@ -12,17 +11,14 @@ import android.text.TextUtils;
 import javax.inject.Inject;
 
 import fr.free.nrw.commons.BuildConfig;
-import fr.free.nrw.commons.category.CategoryDao;
 import fr.free.nrw.commons.data.DBOpenHelper;
 import fr.free.nrw.commons.di.CommonsDaggerContentProvider;
 import timber.log.Timber;
 
-import static android.content.UriMatcher.NO_MATCH;
-import static fr.free.nrw.commons.bookmarks.BookmarkDao.Table.ALL_FIELDS;
-import static fr.free.nrw.commons.bookmarks.BookmarkDao.Table.COLUMN_MEDIA_NAME;
-import static fr.free.nrw.commons.bookmarks.BookmarkDao.Table.TABLE_NAME;
+import static fr.free.nrw.commons.bookmarks.pictures.BookmarkPictureDao.Table.COLUMN_MEDIA_NAME;
+import static fr.free.nrw.commons.bookmarks.pictures.BookmarkPictureDao.Table.TABLE_NAME;
 
-public class BookmarkContentProvider extends CommonsDaggerContentProvider {
+public class BookmarkPictureContentProvider extends CommonsDaggerContentProvider {
 
     private static final String BASE_PATH = "bookmarks";
     public static final Uri BASE_URI = Uri.parse("content://" + BuildConfig.BOOKMARK_AUTHORITY + "/" + BASE_PATH);
@@ -31,7 +27,8 @@ public class BookmarkContentProvider extends CommonsDaggerContentProvider {
         return Uri.parse(BASE_URI.toString() + "/" + name);
     }
 
-    @Inject DBOpenHelper dbOpenHelper;
+    @Inject
+    DBOpenHelper dbOpenHelper;
 
     @Override
     public String getType(@NonNull Uri uri) {
@@ -76,7 +73,7 @@ public class BookmarkContentProvider extends CommonsDaggerContentProvider {
     @Override
     public Uri insert(@NonNull Uri uri, ContentValues contentValues) {
         SQLiteDatabase sqlDB = dbOpenHelper.getWritableDatabase();
-        long id = sqlDB.insert(BookmarkDao.Table.TABLE_NAME, null, contentValues);
+        long id = sqlDB.insert(BookmarkPictureDao.Table.TABLE_NAME, null, contentValues);
         getContext().getContentResolver().notifyChange(uri, null);
         return Uri.parse(BASE_URI + "/" + id);
     }
